backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
Flask==3.0.3
Flask-Cors==5.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
pytest==8.3.2
pytest-flask==1.3.0
python-dotenv==1.0.1
Werkzeug==3.0.4
